backend/requirements.txt
flask
flask-restful
flask-cors
tensorflow
mediapipe
numpy
pandas
opencv-python
python-dotenv
rembg
pillow
onnxruntime